Begin publishing pacts to new broker
We are working on migrating our Pact Broker instance from PaaS to Heroku. As a first step, this PR updates our CI to begin publishing pacts to both instances simultaneously. We (temporarily) use the `continue-on-error: true` option to ensure that CI does not fail if the push to the new instance fails for any reason.