Add rack to runtime dependencies
This dependency is needed at runtime due to the use of rack/utils in Search [1].
I expect this hasn't been caught previously as this gem is typically
used in an application with rack. I found it out by putting together a
standalone script that used this gem and needed to install rack
separately.

[1]: https://github.com/alphagov/gds-api-adapters/blob/e8ba0e5aa9379e6fb8c11d397832e57b7af4226c/lib/gds_api/search.rb#L2