Added a new stub for attempting to delete a missing asset
This is needed for testing how publishing applications react to attempting to delete a missing asset. At the moment they are not handling not found errors, meaning that the state of the publishing application remains out of sync with asset manager and causing other logical issues.

This commit also adds some test coverage for other asset deletion stubs and makes byebug available in the testing runtime.