pin "rack" gem to v2.2
The previous config allowed rack to be bumped to a higher version (e.g. 3)
which introduced errors when running the test [1] for CI and for new users running locally.

This sets the version to no greater than 2.2 as a temporary fix to allow tests to pass, until
updates can be made to fix the breaking changes.

[1] https://gds.slack.com/archives/C03D792LYJG/p1718295709926239?thread_ts=1718203657.261269&cid=C03D792LYJG