Add stub method for getting content that embeds a target
In Whitehall we use this method before all of our Cucumber tests as a default stub. This enables us to make our tests more concise. We can avoid having plumb this stub into various scenarios where the dependend content is not under tests. To do this we set up a stub that allows any content_id (via `anything`) so long as it looks like a uuid. If such a request is made we’ll return back an empty set of results.