Return GraphQL response as `GdsApi::Response`
This is instead of `Hash`.

In Government Frontend, we need some other information from the request
response (e.g. the cache control headers) which are not included when
the response is converted to a hash.

This is not a breaking change as the new object type accepts the same
methods as the hash.