Return GraphQL responses with underscore keys
Publishing API returns GraphQL responses with camel cased keys. However
we use underscored keys in our applications.

Therefore transforming the keys here, so the same does not need to be
duplicated in every consuming application.

Note: the activesupport gem is required, as `deep_transform_keys` (on
Hash) and `underscore` (on String) are both part of Rails, not Ruby.
This avoids needing to rewrite those methods in this application.