Update pact test to reflect that Publishing API publish endpoint is now idempotent
The Publishing API publish endpoint used to return HTTP 409 Conflict error responses if a publishing app attempted to publish a content item which already had a published edition and no draft edition.

These errors are rarely very useful to the Publishing apps as they are usually caused by a duplicate request due to dependency resolution behaviour or Sidekiq's "at least once" guarantee.

We therefore updated Publishing API to return a 200 OK HTTP response is such cases, as the system is still in a perfectly valid state.