Split out publish workflow from CI
In our apps we split the release process into its own workflow. The
workflow only runs if CI has completed on the main branch

In contrast, in our gems the publish process is a final step of the CI
workflow. Though publishing isn't in a separate workflow, it does still
depend on the tests passing

Splitting out the publish process into its own workflow like we do with
releasing apps would provide a few benefits:
- a finer grain look at failires in GitHub's [actions view][] (without
  having to open each individual failed run)
- the ability to trigger a publish manually without going through the
  rest of the CI workflow, should the tests pass but the publish fail,
  for instance
- the ability to add [failure alerts][] to CI and publishing
  independently and with different conditions if useful

[actions view]: https://github.com/alphagov/gds-api-adapters/actions
[failure alerts]: https://github.com/alphagov/govuk-infrastructure/tree/main/.github/actions/report-run-failure