Add support for unpublished editions in `graphql_content_item` method
The `graphql_content_item` method is intended to transform Publishing
API's GraphQL response to make it look like a content item returned from
Content Store. This is so we can implement GraphQL whilst making as few
changes to the frontend applications as possible.

As part of this, we need to handle unpublished editions correctly, by
returning the content of the error in the response, so the frontend
application can handle it.