WIP Expose request Content-Length to API clients
We've been thinking about the issue of the gradually increasing payload
sizes sent from publishing apps to Publishing API.

Here's just one of the places that we could start to surface payload
sizes either back to the client directly or into some kind of
monitoring. The first thought was to put this in Publishing API, but
this option was an interesting idea because it seemed to put the
information in the hands of the publishing app's even more directly.