Remove Publishing API's `graphql_content_item` method
This method previously took a POST request with a GraphQL query and
returned a response similar to Content Store. This has been replaced
with `graphql_live_content_item` that behaves exactly like Content
Store.

Both uses of this method elsewhere in GOV.UK have been removed, so the
method is no longer required.