Pact updates for removal of ?version= in GET content
We've replaced the version param with a content_store param.

This means it's no longer possible to request a superseded edition (so
that pact is removed).

It's still possible to request published content instead of draft (by
querying with content_store=live), or to request draft content
specifically (with content_store=draft).